requirements.txt
alembic==1.4.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-CKEditor==0.4.3
Flask-Cors==3.0.9
Flask-JSGlue==0.3.1
Flask-Mail==0.9.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-SSLify==0.1.5
flask-transfer==0.1.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.4.0
marshmallow-sqlalchemy==0.22.2
mysqlclient==1.4.6
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.19.0
python-editor==1.0.4
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.11
Werkzeug==1.0.0
WTForms==2.2.1
